==> step2.Rmd <==
---
title: "CSCI 3701 Project Step 2"
author: "Paul Friederichsen"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning=FALSE, message=FALSE, error=FALSE)
library(tidyverse)
library(rlist)
library(furrr)
library(knitr)
library(plotly)

future::plan(multiprocess)
chunk2 <- function(x,n) split(x, cut(seq_along(x), n, labels = FALSE)) 
```

[GitHub Repo](https://github.com/floogulinc/csci-3701-project)

## Step 1 recap

My data comes from the [Mathematics Genealogy Project](https://www.genealogy.math.ndsu.nodak.edu/) at NDSU. It is a database of people who have received as doctorate (or similar) in mathematics. This website includes these data points for a given individual:

Their complete name
- Their degrees/dissertations For each university/dissertation:
  - The name of the university 
  - The year(s) in which their degree was awarded
  - The complete title of the dissertation
  - The complete names of their advisors
  
I scrapped all the pages on the site (253772 of them) and then converted them into a dataframe (tibble).

Let's import that now:
```{r}
site.data <- readRDS("sitedata.Rda")
format(object.size(site.data), units="Mb")
```

That process included a good amount of data analysis in itself as I had to handle any pages with odd data that didn't follow what seemed to be the general format. Here are a couple of examples of this from step 1:

<blockquote style="font-size: 14px">

While working on this part I encountered rare instances of odd data. For example, there are some people with no country flag, or just one while they have multiple universities. I additionally found [a case](https://www.genealogy.math.ndsu.nodak.edu/id.php?id=252229) of someone where for a single dissertation there are two flags. Because of this I had to switch from a "country" field in the inner schools tibble to a "countries" field that is a list.

Another extremely rare one that was difficult to figure out was [the instance](https://www.genealogy.math.ndsu.nodak.edu/id.php?id=236498) of "Ph.D. advisor:" instead of the normal "Advisor" in the text where the advisor links are. In [another instance](https://www.genealogy.math.ndsu.nodak.edu/id.php?id=93048) it was "Supervisor" instead. One of [this person's](https://www.genealogy.math.ndsu.nodak.edu/id.php?id=127962) is "Doctoral advisor". I finally generalized the xpath for the advisor `p` tag with it's style, while keeping the search for "Advisor" because some pages with no advisor don't have the same style for that `p` tag.

</blockquote>

Some other issues I ran into were:

- IDs without a record, they just have a blank page that says You have specified an ID that does not exist in the database. Please back up and try again."
- People with multiple degrees/dissertations
- People with two country flags for a single University/degree
- Degrees without a listed advisor, country, or university

Looking at the case of multiple flags, we can find how many instances of those exist.

```{r}
allschools <- map_dfr(site.data$schools, ~ .)
allschools[map_lgl(allschools$countries, ~ length(.) > 1),c(1,2,4)]
```

The second one is an interesting example, [3 flags](https://www.genealogy.math.ndsu.nodak.edu/id.php?id=7824) (for three colleges).

## More data analysis

Lets look at some trends and graphs.

### Number of advisors and students
Let's look at the number of advisors and students people have.

```{r, out.width="100%", fig.height=4}
advisorCounts <- map_int(site.data$advisors, length)

summary(advisorCounts)

plot_ly(y=table(advisorCounts), type="bar") %>% layout(title="Frequency of Advisor Count", xaxis=list(title="Number of Advisors"), yaxis=list(title="Frequency"))
```



```{r, out.width="100%", fig.height=4}
studentCounts <- map_int(site.data$students.id, length)

summary(studentCounts)

plot_ly(y=table(studentCounts)[0:8], type="bar") %>% layout(title="Frequency of Student Count (< 8 students)", xaxis=list(title="Number of Students"), yaxis=list(title="Frequency"))

```

I cut off the X-axis at 7 but there are some people with many more students than that, lets look at those with more than 8 students.

```{r, out.width="100%", fig.height=4}
plot_ly(y=table(studentCounts)[8:max(studentCounts)], type="bar") %>% layout(title="Frequency of Student Count (> 7 students)", xaxis=list(title="Number of Students"), yaxis=list(title="Frequency"))
```

### Dissertations by Country

```{r, out.width="100%", fig.height=4}
filterfreq <- 500
allcountries <- unlist(allschools$countries)
countrycounts <- plyr::count(allcountries)
filteredcountrycounts <- filter(countrycounts, freq >= filterfreq)
filteredcountrycounts$x <- as.character(filteredcountrycounts$x)
filteredcountrycounts <- rbind(filteredcountrycounts, list("Other", sum(countrycounts$freq[countrycounts$freq < filterfreq])))

(ggplot(filteredcountrycounts, aes(x, freq)) + geom_bar(stat="identity") + theme(axis.text.x = element_text(angle=45,hjust=1)) + xlab("Country") + ggtitle("Dissertations by Country")) %>% ggplotly()
```

It looks like the majority of people in the database are from (or studied in) the United States, with another good portion from Germany, France, and the U.K.

Lets look at the same data on a map.

```{r, error=FALSE, warning=FALSE, message=FALSE}
library(countrycode)
countrycountsiso <- countrycounts
countrycountsiso$c <- countrycode(countrycountsiso$x, "country.name", "iso3c")
nacount <- sum(countrycounts$freq[is.na(countrycountsiso$c)])
#countrycountsiso <- countrycountsiso[!is.na(countrycountsiso$c),]
```


```{r, out.width="100%"}
plot_geo() %>%
  add_trace(
    z = countrycountsiso$freq, color = countrycountsiso$freq, colors = 'Blues',
    text = countrycountsiso$x, locations = countrycountsiso$c, marker = list(line = list(color = toRGB("grey"), width = 0.5))) %>%
  colorbar(title = 'Number of Dissertations', type="log") %>%
  layout(
    title = 'Dissertations by Country Map',
    geo = list(
      showcountries = TRUE,
      countrycolor = toRGB("light grey"),
      showframe = FALSE,
      showcoastlines = FALSE,
      projection = list(type = 'Mercator')
    )
  )

```


## Bringing it into a graph library

The next step is to bring the data into some graph database or library. The first step to do so is to get the nodes and edges figured out.

I will create a new datatable of just the ID and names of everyone to be the nodes and a datatable of connections, the edges.

```{r}
nodes <- site.data[c("id", "name")]
```

There will be two types of edges, those representing that the `from` id is a student of the `to` ID and those representing that the `from` ID is an advisor to the `to` ID.

```{r, eval=FALSE}
splitdf <- function(df, n) {
  split(df, (seq(nrow(df))-1) %/% n) 
}

edges.studentOf <- future_map_dfr(splitdf(site.data[,c("id", "advisors")], 10000), unnest, cols = "advisors", .progress = TRUE)
edges.studentOf$type <- "studentOf"
edges.studentOf$advisors <- unlist(edges.studentOf$advisors) # advisors field are all lists of 1 from unnest
colnames(edges.studentOf) <- c("from", "to", "type")

saveRDS(edges.studentOf, "edges-student-of.Rda")

edges.advisorTo <- future_map_dfr(splitdf(site.data[,c("id", "students.id")], 10000), unnest, cols = "students.id", .progress = TRUE)
edges.advisorTo$type <- "advisorTo"
edges.advisorTo$students.id <- unlist(edges.advisorTo$students.id) # students.id field are all lists of 1 from unnest
colnames(edges.advisorTo) <- c("from", "to", "type")

saveRDS(edges.advisorTo, "edges-advisor-to.Rda")

edges <- rbind(edges.studentOf, edges.advisorTo)

saveRDS(edges, "alledges.Rda")
```

Running the operation in parallel really helped with RAM usage and made it fairly quick.

Now lets look at what was created. We should have a simple dataframe with `from`, `to`, and `type`.

```{r}
edges <- readRDS("alledges.Rda")
edges
```

Now let's bring that into the `igraph` library.

```{r}
library(igraph)

net <- graph_from_data_frame(edges, vertices = nodes, directed = TRUE)
net
format(object.size(net), units="Mb")
```

We can also make one with the full set of data for the nodes, but it will be a larger structure.

```{r}
net2 <- graph_from_data_frame(edges, vertices = site.data, directed = TRUE)
net2
format(object.size(net2), units="Mb")
```

So now we have it in `igraph`. This is somewhat of a sanity check in itself since it imported everything fine.

Let's chart some parts of the graph.

Let's chart the first 10 edges (and their nodes) with the `visNetwork` library (they are interactive and can be panned and zoomed).

```{r, out.width="100%"}
edge_attr(net, "label") <- edge_attr(net, "type") # set label to type so we see the edge type

library(visNetwork)

visIgraph(subgraph.edges(net, eids=1:10))

```

Now for a larger example, the first 200 edges.

```{r, out.width="100%"}
visIgraph(subgraph.edges(net, eids=1:200))
```

Let's look at the neighborhood (order 3) of a specific person, that is those within 3 degrees of seperation.

```{r, out.width="100%"}
induced_subgraph(net, unlist(ego(net, nodes = 54719, order = 3))) %>% visIgraph()
```

## What's next

We now run into the issue that none of the verious graph libraries for R are too good at dealing with such a large graph (most of the things I attempt to do with `igraph` seem to take forever or crash R) so in step 3 I will try to get the data into a dedicated graph database, potentially neo4j, and do further analysis and exploration from there.
